# ggml-cpu : add chunking support to mul_mat_id