@max-krasnyansky 2025-02-05
Looks good but seems to eat a few T/S on X-Elite for me.

before

| model                          |       size |     params | backend    | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ------: | ------------: | -------------------: |
| olmoe A1.7B Q4_0               |   3.66 GiB |     6.92 B | CPU        |      10 |         pp128 |        544.64 ± 4.73 |
| olmoe A1.7B Q4_0               |   3.66 GiB |     6.92 B | CPU        |      10 |          tg64 |        112.80 ± 0.85 |

after

| model                          |       size |     params | backend    | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ------: | ------------: | -------------------: |
| olmoe A1.7B Q4_0               |   3.66 GiB |     6.92 B | CPU        |      10 |         pp128 |        534.85 ± 4.62 |
| olmoe A1.7B Q4_0               |   3.66 GiB |     6.92 B | CPU        |      10 |          tg64 |        112.15 ± 1.33 |

will recheck again tomorrow