@slaren 2025-02-05
I have changed it to use a different atomic chunk counter for each matrix, which should remove the need of most of the synchronization.

With 13900k:
| Model             |   Threads | Test   |   t/s master |   t/s sl/mmid-cpu-perf |   Speedup |
|:------------------|----------:|:-------|-------------:|-----------------------:|----------:|
| llama 8x7B Q3_K_S |         8 | pp64   |        23.36 |                  27.11 |      1.16 |
| llama 8x7B Q3_K_S |         8 | tg32   |        12.30 |                  14.13 |      1.15 |
| llama 8x7B Q3_K_S |        16 | pp64   |        18.42 |                  35.99 |      1.95 |
| llama 8x7B Q3_K_S |        16 | tg32   |        10.42 |                  15.04 |      1.44 |
| llama 8x7B Q3_K_S |        24 | pp64   |        26.21 |                  41.35 |      1.58 |
| llama 8x7B Q3_K_S |        24 | tg32   |        11.23 |                  13.13 |      1.17 |
| llama 8x7B Q3_K_S |        32 | pp64   |        32.68 |                  40.74 |      1.25 |
| llama 8x7B Q3_K_S |        32 | tg32   |        10.99 |                  11.46 |      1.04 |