@max-krasnyansky 2025-02-05
> The chunk size could probably be tuned for different processors, it has a significant effect on performance. From what I could find, the X-Elite does not actually have heterogeneous cores, so I wouldn't expect an improvement from these changes, and increasing the chunk size should reduce the performance loss from the additional synchronization that this change requires.

Yep. X-Elite has uniform cores, so I wasn't expecting the gain just wanted to make sure we don't see regressions.
I will check your latest update a bit later and also test on S24 and S25 with het.cores.