@max-krasnyansky 2025-02-09
> @max-krasnyansky I realized that Q4_0 does not use the same mat mul function due to the aarch64 conversion. So if you test this again, do not use a Q4_0 model, it is using the same code as on master and any differences should be just noise.

I'll try Q2_K or Q4_K shortly