@max-krasnyansky 2025-02-09
Here is another set of runs with Q4_K on X-Elite.
Looks like some further tunning will be needed. 
master is definitely better, chunk size 64 helps with fewer threads but hurts with more threads. 

I'll play with it some more tomorrow. 
btw I realized the regular matmul has the same issue. I vaguely remember that original chunking support caused regression
but with `Q4_0_4_8` we got a big bump and I forgot about it.  


`master\bin\llama-bench.exe -m C:\Users\maxk\src\gguf\OLMoE-1B-7B-0924.Q4_K.gguf -p 128 -n 0 -t 2,4,6,8,10 -ngl 0 --delay 15`

| model                          |       size |     params | backend    | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ------: | ------------: | -------------------: |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       2 |         pp128 |         56.82 ± 1.04 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       4 |         pp128 |        110.64 ± 0.43 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       6 |         pp128 |        163.05 ± 0.81 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       8 |         pp128 |        214.60 ± 0.66 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |      10 |         pp128 |        263.76 ± 1.97 |

build: 55ac8c77 (4675)

chunk-size-16
`mmid-cs16\bin\llama-bench.exe -m C:\Users\maxk\src\gguf\OLMoE-1B-7B-0924.Q4_K.gguf -p 128 -n 0 -t 2,4,6,8,10 -ngl 0 --delay 15`

| model                          |       size |     params | backend    | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ------: | ------------: | -------------------: |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       2 |         pp128 |         49.48 ± 0.44 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       4 |         pp128 |        100.48 ± 0.30 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       6 |         pp128 |        152.68 ± 0.25 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       8 |         pp128 |        203.39 ± 0.95 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |      10 |         pp128 |        250.23 ± 2.08 |

build: 45cede2f (4629)

chunk-size-64
`mmid-cs64\bin\llama-bench.exe -m C:\Users\maxk\src\gguf\OLMoE-1B-7B-0924.Q4_K.gguf -p 128 -n 0 -t 2,4,6,8,10 -ngl 0 --delay 15`

| model                          |       size |     params | backend    | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ------: | ------------: | -------------------: |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       2 |         pp128 |         57.30 ± 1.37 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       4 |         pp128 |        111.17 ± 0.30 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       6 |         pp128 |        164.98 ± 0.56 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |       8 |         pp128 |        199.64 ± 1.03 |
| olmoe A1.7B Q4_K - Medium      |   3.92 GiB |     6.92 B | CPU        |      10 |         pp128 |        219.92 ± 1.55 |

build: 45cede2f (4629)