@slaren 2025-02-09
Thanks for testing. It's surprising that increasing the chunk size decreases performance, I would expect that a larger chunk size should make the behavior closer to the previous behavior. I am not sure what could cause this, maybe it is some cache effect that I am missing?

I don't have the hardware to test this, so all I can do is disable this for ARM. On M3 Max the change is mostly neutral, with better performance above 12 threads (but still worse than 12).

| CPU                      | Model              |   Threads | Test   |   t/s master |   t/s sl/mmid-cpu-perf |   Speedup |
|:-------------------------|:-------------------|----------:|:-------|-------------:|-----------------------:|----------:|
|  Apple M3 Max | olmoe A1.7B Q4_K_M |         2 | pp128  |        97.00 |                  92.91 |      0.96 |
|  Apple M3 Max | olmoe A1.7B Q4_K_M |         4 | pp128  |       181.95 |                 178.70 |      0.98 |
|  Apple M3 Max | olmoe A1.7B Q4_K_M |         8 | pp128  |       320.09 |                 317.74 |      0.99 |
|  Apple M3 Max | olmoe A1.7B Q4_K_M |        12 | pp128  |       386.02 |                 381.51 |      0.99 |
|  Apple M3 Max | olmoe A1.7B Q4_K_M |        13 | pp128  |       244.26 |                 355.65 |      1.46 |
|  Apple M3 Max | olmoe A1.7B Q4_K_M |        14 | pp128  |       241.09 |                 355.24 |      1.47 |
|  Apple M3 Max | olmoe A1.7B Q4_K_M |        15 | pp128  |       225.00 |                 327.30 |      1.45 |
|  Apple M3 Max | olmoe A1.7B Q4_K_M |        16 | pp128  |       184.14 |                 260.59 |      1.42 |